+++ user/ulib.zig
@@ -8,7 +8,7 @@ pub export fn start(argc: i32, argv: [*][*:0]u8) void {
 }
 
 pub export fn strcpy(dst: [*]u8, src: [*:0]const u8) [*:0]u8 {
-    var i: usize = 0;
+    var i: u64 = 0;
     while (true) : (i += 1) {
         dst[i] = src[i];
         if (src[i] == 0) {
@@ -19,21 +19,21 @@ pub export fn strcpy(dst: [*]u8, src: [*:0]const u8) [*:0]u8 {
 }
 
 pub export fn strcmp(p: [*:0]const u8, q: [*:0]const u8) i32 {
-    var i: usize = 0;
+    var i: u64 = 0;
     while (p[i] != 0 and p[i] == q[i]) : (i += 1) {}
     return @as(i32, p[i]) - @as(i32, q[i]);
 }
 
 pub export fn strlen(s: [*:0]const u8) u32 {
-    var n: usize = 0;
+    var n: u64 = 0;
     while (s[n] != 0) : (n += 1) {}
     return @intCast(n);
 }
 
 pub export fn memset(dst: [*]u8, c: i32, n: u32) ?*anyopaque {
-    var i: usize = 0;
+    var i: u64 = 0;
     const ch: u8 = @intCast(c & 0xff);
-    const nn: usize = @intCast(n);
+    const nn: u64 = @intCast(n);
     while (i < nn) : (i += 1) {
         dst[i] = ch;
     }
@@ -41,7 +41,7 @@ pub export fn memset(dst: [*]u8, c: i32, n: u32) ?*anyopaque {
 }
 
 pub export fn strchr(s: [*:0]const u8, c: u8) ?[*:0]u8 {
-    var i: usize = 0;
+    var i: u64 = 0;
     while (s[i] != 0) : (i += 1) {
         if (s[i] == c) {
             return @constCast(s + i);
@@ -79,7 +79,7 @@ pub export fn stat(n: [*:0]const u8, st: *xv6.Stat) i32 {
 }
 
 pub export fn atoi(s: [*:0]const u8) i32 {
-    var i: usize = 0;
+    var i: u64 = 0;
     var n: i32 = 0;
     while (s[i] >= '0' and s[i] <= '9') : (i += 1) {
         n = n * 10 + (@as(i32, s[i]) - '0');
@@ -91,14 +91,14 @@ pub export fn memmove(vdst: [*]u8, vsrc: [*]const u8, n: i32) ?*anyopaque {
     if (@intFromPtr(vsrc) > @intFromPtr(vdst)) {
         var i: i32 = 0;
         while (i < n) : (i += 1) {
-            const k: usize = @intCast(i);
+            const k: u64 = @intCast(i);
             vdst[k] = vsrc[k];
         }
     } else {
         var i: i32 = n;
         while (i > 0) {
             i -= 1;
-            const k: usize = @intCast(i);
+            const k: u64 = @intCast(i);
             vdst[k] = vsrc[k];
         }
     }
@@ -108,8 +108,8 @@ pub export fn memmove(vdst: [*]u8, vsrc: [*]const u8, n: i32) ?*anyopaque {
 pub export fn memcmp(s1: ?*const anyopaque, s2: ?*const anyopaque, n: u32) i32 {
     const p1: [*]const u8 = @ptrCast(s1.?);
     const p2: [*]const u8 = @ptrCast(s2.?);
-    var i: usize = 0;
-    const nn: usize = @intCast(n);
+    var i: u64 = 0;
+    const nn: u64 = @intCast(n);
     while (i < nn) : (i += 1) {
         if (p1[i] != p2[i]) {
             return @as(i32, p1[i]) - @as(i32, p2[i]);
